Changed:
-ParaBlit enums now use 'enum class' to scope enums and fix related warnings.
-Most ParaBlit enums follow a new naming convention, as they are now scoped.

Fixed:

-Lots of warnings.
-Swapchain was selecting Vulkan present modes using ParaBlit present modes without proper translation.

+++ VKR/VKRClient/QuickIO.cpp
@@ -19,7 +19,7 @@ namespace QIO
 			*data = buf;
 
 			inStream.close();
-			printf_s("Successfully read %i bytes of data from file at %s \n", *fileSize, path);
+			printf_s("Successfully read %u bytes of data from file at %s \n", (unsigned int)*fileSize, path);
 			return true;
 		}
 		else
@@ -44,7 +44,7 @@ namespace QIO
 			*data = buf;
 
 			inStream.close();
-			printf_s("Successfully read %i bytes of data from file at %s \n", *fileSize, path);
+			printf_s("Successfully read %u bytes of data from file at %s \n", (unsigned int)*fileSize, path);
 			return true;
 		}
 		else
